refactor: TawkTo logic to plugin
