+++ src/pages/AboutUs.vue
@@ -1,6 +1,5 @@
 <!-- urbverde-ui/src/pages/AboutUs.vue -->
 <template>
-  <TawkTo />
   <InstitutionalLayout
     active-item="sobre"
     page-title="Quem somos"
@@ -132,7 +131,6 @@
 <script>
 import InstitutionalLayout from '@/layouts/InstitutionialLayout.vue';
 import MemberCard from '@/components/institutional_pages/MemberCard.vue';
-import TawkTo from '@/components/homepage/TawkTo.vue';
 import { useHead } from '@vueuse/head';
 
 // Arquivo de imports de images e dicionários para times e membros
@@ -143,7 +141,6 @@ export default {
   components: {
     InstitutionalLayout,
     MemberCard,
-    TawkTo,
   },
 
   setup() {
