+++ src/pages/WorkTogether.vue
@@ -1,6 +1,5 @@
 <!-- urbverde-ui/src/pages/WorkTogether.vue -->
 <template>
-  <TawkTo />
   <InstitutionalLayout
     active-item="parceiro"
     page-title="Seja parceiro"
@@ -53,7 +52,6 @@
 <script>
 import InstitutionalLayout from '@/layouts/InstitutionialLayout.vue';
 import InstitutionalContactWrapper from '@/components/institutional_pages/InstitutionalContactWrapper.vue';
-import TawkTo from '@/components/homepage/TawkTo.vue';
 import { useHead } from '@vueuse/head';
 
 export default {
@@ -61,7 +59,6 @@ export default {
   components: {
     InstitutionalLayout,
     InstitutionalContactWrapper,
-    TawkTo,
   },
 
   setup() {
